requirements_gui.txt
matplotlib>=3.5.0
psycopg2-binary>=2.9.0
numpy>=1.21.0
psutil>=5.8.0
GPUtil>=1.4.0